requirements.txt
anyio==3.7.1
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
chardet==4.0.0
charset-normalizer==3.4.1
dataclasses==0.6
deprecation==2.1.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-storage-supabase==0.1.0
djangorestframework==3.15.2
gotrue==0.2.0
gunicorn==23.0.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==2.10
iniconfig==2.0.0
legacy-cgi==2.6.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
postgrest-py==0.5.0
psycopg2==2.9.10
py==1.11.0
pytest==6.2.5
python-dateutil==2.9.0.post0
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
supabase==0.0.3
toml==0.10.2
urllib3==1.26.20
websockets==9.1
whitenoise==6.8.2
